* Correctly register MIME type. * Don't try to associate with the MIME type for now, because it doesn't   really work yet. * Make the weight entry box allow floats. -Samuel

+++ trunk/src/consumerdb.py
@@ -67,7 +67,7 @@ class ConsumerDB(saxutils.DefaultHandler):
             self.cur_day.add_food(self.cur_food)
             
         elif name == "weight":
-            self.cur_day.set_weight(atoi(self.cur_val))
+            self.cur_day.set_weight(atof(self.cur_val))
             
         elif name == "quantity":
             self.cur_food.quantity = atof(self.cur_val)
